Moving history from widget to screens
